pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.1"
python-dotenv = "^0.15.0"
flask = "^1.1.2"
bs4 = "^0.0.1"
youtube_dl = "^2021.6.6"
discord-ui = "^5.1.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
